app/requirements/base.txt
django==3.1
django-model-utils==4.0.0  # Набор полезных базовых классов и утилит для Django
psycopg2-binary==2.8.6
gunicorn==20.0.4
elasticsearch-dsl==7.3.0
pydantic==1.8.1
redis==3.5.3
sqlparse==0.4.1
requests==2.25.1